status codes added

+++ CinemaApi/Controllers/MoviesController.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using System.Threading.Tasks;
 using CinemaApi.Data;
 using CinemaApi.Models;
+using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 
 // For more information on enabling Web API for empty projects, visit https://go.microsoft.com/fwlink/?LinkID=397860
@@ -23,44 +24,56 @@ namespace CinemaApi.Controllers
 
         // GET: api/<MoviesController>
         [HttpGet]
-        public IEnumerable<Movie> Get()
+        public IActionResult Get()
         {
-            return _dbContext.Movies;
+            return Ok(_dbContext.Movies);
         }
 
         // GET api/<MoviesController>/5
         [HttpGet("{id}")]
-        public Movie Get(int id)
+        public IActionResult Get(int id)
         {
             var movie = _dbContext.Movies.Find(id);
-            return movie;
+            if (movie == null) return NotFound("Movie is not found");
+            return Ok(movie);
         }
 
         // POST api/<MoviesController>
         [HttpPost]
-        public void Post([FromBody] Movie movie)
+        public IActionResult Post([FromBody] Movie movie)
         {
             _dbContext.Movies.Add(movie);
             _dbContext.SaveChanges();
+            return StatusCode(StatusCodes.Status201Created);
         }
 
         // PUT api/<MoviesController>/5
         [HttpPut("{id}")]
-        public void Put(int id, [FromBody] Movie movie)
+        public IActionResult Put(int id, [FromBody] Movie newMovie)
         {
-            var item = _dbContext.Movies.Find(id);
-            item.Name = movie.Name;
-            item.Language = movie.Language;
-            _dbContext.SaveChanges();
+            var movie = _dbContext.Movies.Find(id);
+            if (movie == null) return NotFound("Movie is not found"); 
+            else
+            {
+                movie.Name = newMovie.Name;
+                movie.Language = newMovie.Language;
+                _dbContext.SaveChanges();
+                return Ok("Movie was updated");
+            }
         }
 
         // DELETE api/<MoviesController>/5
         [HttpDelete("{id}")]
-        public void Delete(int id)
+        public IActionResult Delete(int id)
         {
             var movie = _dbContext.Movies.Find(id);
-            _dbContext.Movies.Remove(movie);
-            _dbContext.SaveChanges();
+            if (movie == null) return NotFound("Movie is not found");
+            else
+            {
+                _dbContext.Movies.Remove(movie);
+                _dbContext.SaveChanges();
+                return Ok("Movie was removed
+            }
         }
     }
 }
